requirements.txt
pillow
request 
opencv-python 
numpy 
pytesseract 
python-dotenv
